src/main/java/com/javacore/jsonparser/MapSorter.java
package com.javacore.jsonparser;

import java.util.Comparator;
import java.util.LinkedHashMap;
import java.util.Map;
import java.util.stream.Collectors;

public class MapSorter {
    public Map<String, Integer> sortMapByValueInReverseOrder(Map<String, Integer> inputMap) {
        return inputMap.entrySet().stream()
                .sorted(Map.Entry.comparingByValue(Comparator.reverseOrder()))
                .collect(
                        Collectors.toMap(
                                Map.Entry::getKey,
                                Map.Entry::getValue,
                                (oldValue, newValue) -> oldValue,
                                LinkedHashMap::new
                        )
                );
    }
}
